fix: always use %s in assert format string
We don't know enough information to do better.
For abort/throw, %s is the right one but it might not be so
for custom functions. If the format string is converted to unicode
with the L prefix, formatting won't work correctly with ANSI
strings. There, %hs is required with MSVC.